test(python): Move credential provider tests to separate file (#19639)
